Whenever Auth_Container_DB attempts to change the database make sure the connection/DB object exist and are correct before attempting to use the DB quoting features
git-svn-id: http://svn.php.net/repository/pear/packages/Auth/trunk@207647 c90b9560-bf6c-de11-be94-00142212c4b1

+++ Container/DB.php
@@ -396,6 +396,11 @@ class Auth_Container_DB extends Auth_Container
      */
     function removeUser($username)
     {
+        $err = $this->_prepare();
+        if ($err !== true) {
+            return PEAR::raiseError($err->getMessage(), $err->getCode());
+        }
+
         $query = sprintf("DELETE FROM %s WHERE %s = '%s'",
                          $this->options['table'],
                          $this->options['usernamecol'],
@@ -419,6 +424,11 @@ class Auth_Container_DB extends Auth_Container
      */
     function changePassword($username, $password)
     {
+        $err = $this->_prepare();
+        if ($err !== true) {
+            return PEAR::raiseError($err->getMessage(), $err->getCode());
+        }
+
         if (isset($this->options['cryptType']) && $this->options['cryptType'] == 'none') {
             $cryptFunction = 'strval';
         } elseif (isset($this->options['cryptType']) && function_exists($this->options['cryptType'])) {
